requirements/requirements.txt
alembic==1.7.5
anyio==3.4.0
apache-airflow==2.2.3
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-google==6.4.0
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
argcomplete==1.12.3
async-generator==1.10
attrs==20.3.0
Authlib==0.15.5
Babel==2.9.1
beautifulsoup4==4.10.0
blinker==1.4
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.4
cattrs==1.0.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==7.1.2
clickclick==20.10.2
cloudevents==1.2.0
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
contextvars==2.4
croniter==1.0.15
cryptography==36.0.1
dataclasses==0.8
DateTime==4.4
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
diagrams==0.21.1
dill==0.3.1.1
distlib==0.3.4
dnspython==2.1.0
docopt==0.6.2
docutils==0.16
email-validator==1.1.3
filelock==3.4.1
finvizfinance==0.12.2
Flask==1.1.4
Flask-AppBuilder==3.4.1
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
functions-framework==3.0.0
future==0.18.2
google-ads==12.0.0
google-api-core==1.31.5
google-api-python-client==1.12.10
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
google-cloud-aiplatform==1.11.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-automl==2.7.1
google-cloud-bigquery==2.34.1
google-cloud-bigquery-datatransfer==3.6.0
google-cloud-bigquery-storage==2.12.0
google-cloud-bigtable==1.7.0
google-cloud-build==3.8.0
google-cloud-container==1.0.1
google-cloud-core==1.7.2
google-cloud-datacatalog==3.7.0
google-cloud-dataproc==3.3.0
google-cloud-dataproc-metastore==1.4.0
google-cloud-dlp==1.0.0
google-cloud-kms==2.11.0
google-cloud-language==1.3.0
google-cloud-logging==2.7.0
google-cloud-memcache==1.0.0
google-cloud-monitoring==2.9.0
google-cloud-orchestration-airflow==1.3.0
google-cloud-os-login==2.6.0
google-cloud-pubsub==2.10.0
google-cloud-redis==2.7.0
google-cloud-secret-manager==1.0.0
google-cloud-spanner==1.19.1
google-cloud-speech==1.3.2
google-cloud-storage==1.44.0
google-cloud-tasks==2.8.0
google-cloud-texttospeech==1.0.1
google-cloud-translate==1.7.0
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-cloud-workflows==1.6.0
google-crc32c==1.3.0
google-resumable-media==2.3.1
googleapis-common-protos==1.55.0
graphviz==0.19.1
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-gcp==0.2.2
grpcio-status==1.44.0
gunicorn==20.1.0
h11==0.12.0
html5lib==1.1
httpcore==0.13.7
httplib2==0.20.4
httpx==0.19.0
identify==2.4.4
idna==3.3
immutables==0.16
importlib-metadata==4.8.3
importlib-resources==5.2.3
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
ipython-genutils==0.2.0
iso8601==1.0.2
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
json-merge-patch==0.2
jsonschema==3.2.0
jupyter-core==4.9.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.8.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
multitasking==0.0.10
nbformat==5.1.3
nodeenv==1.6.0
nox==2020.12.31
numpy==1.19.5
oauthlib==3.2.0
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==21.3
pandas==1.1.5
pandas-gbq==0.14.1
pendulum==2.1.2
pep562==1.1
pipreqs==0.4.11
platformdirs==2.4.0
pluggy==1.0.0
praw==7.5.0
prawcore==2.3.0
pre-commit==2.17.0
prison==0.2.1
proto-plus==1.20.3
protobuf==3.19.4
psaw==0.1.0
psutil==5.8.0
psycopg2==2.7.7
py==1.11.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydata-google-auth==1.3.0
Pygments==2.10.0
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.0
PySocks==1.7.1
pytest==7.0.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==10.16.0
rsa==4.8
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
snscrape==0.3.4
soupsieve==2.3.1
SQLAlchemy==1.3.24
sqlalchemy-bigquery==1.4.0
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.9
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
traitlets==4.3.3
typing==3.7.4.3
typing_extensions==4.1.1
unicodecsv==0.14.1
update-checker==0.18.0
uritemplate==3.0.1
urllib3==1.26.9
virtualenv==20.13.2
watchdog==1.0.2
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==1.0.1
WTForms==2.3.3
yarg==0.1.9
yfinance==0.1.70
zipp==3.6.0
zope.interface==5.4.0
